Update build for Android build tools 26 and build plugin 3.0.1.
This fails to build/link, apparently because of some dependencies of nunaliit2-couch-client.

Namely, it adds dependencies to commons-logging, httpclient, and json, which conflict with versions of the libraries built-in to the Android platform. This shouldn't cause build problems, though. It also references the packages javax.servlet.http and javax.servlet, which are unavailable on Android. I think that is more likely to cause the build problems.